+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---------------
 - add support for Factory creatures in Horn of the Abyss (issue #2)
 - add support for Sleepkeeper artifact in Horn of the Abyss
+- fix name of Vial of Dragon Blood artifact in Horn of the Abyss
 
 
 1.7, 2023-08-06